+++ source/config/configSpec.py
@@ -60,6 +60,8 @@ schemaVersion = integer(min=0, default={latestSchemaVersion})
 	whiteNoiseVolume = integer(default=0, min=0, max=100)
 	soundSplitState = integer(default=0)
 	includedSoundSplitModes = int_list(default=list(0, 2, 3))
+	applicationsSoundVolume = integer(default=100, min=0, max=100)
+	ApplicationsVolumeMode = featureFlag(optionsEnum="AppsVolumeAdjusterFlag", behaviorOfDefault="DISABLED")
 
 # Braille settings
 [braille]