+++ user_docs/en/changes.md
@@ -11,6 +11,8 @@
 * Added support for Unicode Normalization to speech and braille output. (#16466 @LeonarddeR).
   * This can be of help when reading characters that are unknown to a particular speech synthesizer or braille table and which have a compatible alternative, like the bold and italic characters commonly uses on social media.
   * You can enable this functionality for both speech and braille in their respective settings categories in the NVDA Settings dialog.
+* The volume of the other applications can be adjusted by ``NVDA+alt+pageUp`` and ``NVDA+alt+pageDown``. (#16052, @mltony)
+* The sound of the other applications can be muted with ``NVDA+alt+delete``. (#16052, @mltony)
 
 ### Changes